Control NUMA allocation policy of zsmalloc and ZRAM
This patch adds the numa_node module parameter to zsmalloc and zram
to (optionally) control the NUMA node where these modules allocate the
compressed data and any intermediate buffers that are passed to the
compression algorithm.

If the module parameter is not passed (or if it is < 0) the behavior
of zsmalloc and zram remains unchanged.

If numa_node is passed, the custom allocations use  __GFP_THISNODE
to prevent silent failures of the requested NUMA policy.

NOTE: This patch is experimental and NOT intended to run in
systems with HIGHMEM enabled.

Signed-off-by: Dimitrios Palyvos-Giannas <dimitrios.palyvos@zptcorp.com>